Attendance/requirements.txt

flask
requests
insightface
opencv-python
numpy
pandas
onnxruntime-gpu
mysql-connector-python
Werkzeug
